+++ lognflow/lognflow.py
@@ -393,7 +393,7 @@ class lognflow:
         if(print_text):
             if(log_time_stamp):
                 print(f'T:{self.time_stamp:>6.6f}| ', end='')
-            print(to_be_logged)
+            print(to_be_logged, end = end)
                 
         if ( (not (log_name in self._loggers_dict)) or new_file):
             self._log_text_handler(log_name, 
@@ -412,7 +412,10 @@ class lognflow:
         else:
             _tolog = str(to_be_logged)
             _logger.append(_tolog)
-        if(_logger[-1][-1] != end):
+        if(len(_logger[-1]) > 0):
+            if(_logger[-1][-1] != end):
+                _logger.append(end)
+        else:
             _logger.append(end)
         log_size = 0
         for _logger_el in _logger:
@@ -669,6 +672,7 @@ class lognflow:
     def log_multichannel_by_subplots(self, 
         parameter_name: str, 
         parameter_value: np.ndarray,
+        frame_shape = None,
         image_format='jpeg', 
         dpi=1200, 
         time_tag: bool = None,
@@ -693,13 +697,18 @@ class lognflow:
         time_tag = self.time_tag if (time_tag is None) else time_tag
             
         n_r, n_c, n_ch = parameter_value.shape
-        n_ch_sq = int(np.ceil(n_ch ** 0.5))
-        _, ax = plt.subplots(n_ch_sq,n_ch_sq)
+        
+        if(frame_shape is None):
+            n_ch_sq = int(np.ceil(n_ch ** 0.5))
+            n_ch_r, n_ch_c = (n_ch_sq, n_ch_sq)
+        else:
+            n_ch_r, n_ch_c = frame_shape
+        _, ax = plt.subplots(n_ch_r,n_ch_c)
         if(remove_axis_ticks):
             plt.setp(ax, xticks=[], yticks=[])
-        for rcnt in range(n_ch_sq):
-            for ccnt in range(n_ch_sq):
-                im = parameter_value[:,:, ccnt + rcnt * n_ch_sq]
+        for rcnt in range(n_ch_r):
+            for ccnt in range(n_ch_c):
+                im = parameter_value[:,:, ccnt + rcnt * n_ch_c]
                 im_ch = ax[rcnt, ccnt].imshow(im, **kwargs)
                 if(add_colorbar):
                     self.add_colorbar(im_ch)
@@ -880,7 +889,8 @@ class lognflow:
             n_r, n_c = parameter_value.shape
             fig = plt.figure()
             ax = fig.add_subplot(111, projection='3d')
-            X, Y = np.meshgrid(np.arange(n_r), np.arange(n_c))
+            X, Y = np.meshgrid(np.arange(n_r, dtype='int'), 
+                               np.arange(n_c, dtype='int'))
             ax.plot_surface(X, Y, parameter_value, **kwargs)
             fpath = self.log_plt(
                 parameter_name = parameter_name, 
@@ -1043,15 +1053,92 @@ class lognflow:
                 canv = np.zeros((n_imgs, new_n_R, new_n_C, 3),
                                  dtype = stack.dtype)
             used_ch_cnt = 0
-
-            stack[:,   :1      ] = borders
-            stack[:,   : ,   :1] = borders
-            stack[:, -1:       ] = borders
-            stack[:,   : , -1: ] = borders
+            if(borders is not None):
+                stack[:,   :1      ] = borders
+                stack[:,   : ,   :1] = borders
+                stack[:, -1:       ] = borders
+                stack[:,   : , -1: ] = borders
             
             for rcnt in range(square_side):
                 for ccnt in range(square_side):
                     ch_cnt = rcnt + square_side*ccnt
+                    if (ch_cnt<n_ch):
+                        canv[:, rcnt*n_R: (rcnt + 1)*n_R,
+                                ccnt*n_C: (ccnt + 1)*n_C] = \
+                            stack[..., used_ch_cnt]
+                        used_ch_cnt += 1
+        else:
+            return None
+        return canv
+
+    def multichannel_to_frame(self, stack, 
+                              frame_shape : tuple = None, borders = 0):
+        """ turn a stack of multi-channel images into a frame of images
+            This is very useful when lots of images need to be tiled
+            against each other.
+        
+            :param stack: np.ndarray
+                    It must have the shape of either
+                    n_r x n_c x n_ch
+                    n_r x n_c x  3  x n_ch
+                    n_f x n_r x n_c x n_ch
+                    n_f x n_r x n_c x  3  x n_ch
+                    
+                In both cases n_ch will be turned into a square tile
+                Remember if you have N images to put into a square, the input
+                shape should be 1 x n_r x n_c x N
+            :param frame_shape: tuple
+                The shape of the frame to put n_rows and n_colmnss of images
+                close to each other to form a rectangle of image.
+            :param borders: literal or np.inf or np.nan
+                When plotting images with matplotlib.pyplot.imshow, there
+                needs to be a border between them. This is the value for the 
+                border elements.
+                
+            output
+            ---------
+                Since we have N channels to be laid into a square, the side
+                length woul be ceil(N**0.5)
+                it produces an np.array of shape n_f x n_r * S x n_c * S or
+                n_f x n_r * S x n_c * S x 3 in case of RGB input.
+        """
+        if(len(stack.shape) == 4):
+            if(stack.shape[3] == 3):
+                stack = np.array([stack])
+        if(len(stack.shape) == 3):
+            stack = np.array([stack])
+        
+        if((len(stack.shape) == 4) | (len(stack.shape) == 5)):
+            if(len(stack.shape) == 4):
+                n_imgs, n_R, n_C, n_ch = stack.shape
+            if(len(stack.shape) == 5):
+                n_imgs, n_R, n_C, is_rgb, n_ch = stack.shape
+                if(is_rgb != 3):
+                    return None
+            if(frame_shape is None):
+                square_side = int(np.ceil(np.sqrt(n_ch)))
+                frame_n_r, frame_n_c = (square_side, square_side)
+            else:
+                frame_n_r, frame_n_c = frame_shape
+            
+            new_n_R = n_R * frame_n_r
+            new_n_C = n_C * frame_n_c
+            if(len(stack.shape) == 4):
+                canv = np.zeros((n_imgs, new_n_R, new_n_C), 
+                                dtype = stack.dtype)
+            if(len(stack.shape) == 5):
+                canv = np.zeros((n_imgs, new_n_R, new_n_C, 3),
+                                 dtype = stack.dtype)
+            used_ch_cnt = 0
+            if(borders is not None):
+                stack[:,   :1      ] = borders
+                stack[:,   : ,   :1] = borders
+                stack[:, -1:       ] = borders
+                stack[:,   : , -1: ] = borders
+            
+            for rcnt in range(frame_n_r):
+                for ccnt in range(frame_n_c):
+                    ch_cnt = rcnt + frame_n_c*ccnt
                     if (ch_cnt<n_ch):
                         canv[:, rcnt*n_R: (rcnt + 1)*n_R,
                                 ccnt*n_C: (ccnt + 1)*n_C] = \
@@ -1071,7 +1158,7 @@ class lognflow:
             else:
                 canv = stack
         if((len(stack.shape) == 4) | (len(stack.shape) == 5)):
-            canv = self.multichannel_to_square(stack, borders = borders)
+            canv = self.multichannel_to_frame(stack, borders = borders)
         return canv
     
     def prepare_stack_of_images(self, 
@@ -1361,239 +1448,6 @@ class lognflow:
                  writer = animation.PillowWriter(fps=int(1000/interval)))
         return fpath
 
-    def get_text(self, log_name='main_log'):
-        """ get text log files
-            Given the log_name, this function returns the text therein.
-
-            Parameters
-            ----------
-            :param log_name:
-                the log name. If not given then it is the main log.
-        """
-        flist = list(self.log_dir.glob(f'{log_name}*.txt'))
-        flist.sort()
-        n_files = len(flist)
-        if (n_files>0):
-            txt = []
-            for fcnt in range(n_files):
-                with open(flist[fcnt]) as f_txt:
-                    txt.append(f_txt.readlines())
-            if(n_files == 1):
-                txt = txt[0]
-            return txt
-
-    def get_single(self, var_name, single_shot_index = -1, 
-                     suffix = '.np*'):
-        """ get a single variable
-            return the value of a saved variable.
-
-            :param var_name:
-                variable name
-            :param single_shot_index:
-                If there are many snapshots of a variable, this input can
-                limit the returned to a set of indices.
-            :param suffix:
-                If there are different suffixes availble for a variable
-                this input needs to be set. npy, npz, mat, and torch are
-                supported.
-                
-            .. note::
-                when reading a MATLAB file, the output is a dictionary.
-        """
-        var_name = var_name.replace('\t', '\\t').replace('\n', '\\n')\
-            .replace('\r', '\\r').replace('\b', '\\b')
-        
-        suffix = suffix.strip('.')
-        assert single_shot_index == int(single_shot_index), \
-                    f'single_shot_index {single_shot_index} must be an integer'
-        flist = []            
-        if((self.log_dir / var_name).is_file()):
-            flist = [self.log_dir / var_name]
-        elif((self.log_dir / f'{var_name}.{suffix}').is_file()):
-            flist = [self.log_dir / f'{var_name}.{suffix}']
-        else:
-            _var_name = (self.log_dir / var_name).name
-            _var_dir = (self.log_dir / var_name).parent
-            flist = list(_var_dir.glob(f'{_var_name}*.{suffix}'))
-            if(len(flist) == 0):
-                flist = list(_var_dir.glob(f'{_var_name}*.*'))
-                if(len(flist) > 0):
-                    # self.log_text(None, 
-                    #     'Can not find the file with the given suffix, '\
-                    #     +'but found some with a different suffix, '\
-                    #     +f'one file is: {flist[single_shot_index]}')
-                    pass
-                    
-        if(len(flist) > 0):
-            flist.sort()
-        else:
-            var_dir = self.log_dir / var_name
-            if(var_dir.is_dir()):
-                flist = list(var_dir.glob('*.*'))
-            if(len(flist) > 0):
-                flist.sort()
-            else:
-                # self.log_text(None, 'No such variable')
-                return
-        var_path = flist[single_shot_index]
-                
-        if(var_path.is_file()):
-            # self.log_text(None, f'Loading {var_path}')
-            if(var_path.suffix == '.npz'):
-                buf = np.load(var_path)
-                try:
-                    time_array = buf['time_array']
-                    n_logs = (time_array > 0).sum()
-                    time_array = time_array[:n_logs]
-                    data_array = buf['data_array']
-                    data_array = data_array[:n_logs]
-                    return((time_array, data_array))
-                except:
-                    return buf
-            if(var_path.suffix == '.npy'):
-                return(np.load(var_path))
-            if(var_path.suffix == '.mat'):
-                return(loadmat(var_path))
-            if(var_path.suffix == '.txt'):
-                with open(var_path) as f_txt:
-                    return(f_txt.read())
-            if((var_path.suffix == '.tif') | (var_path.suffix == '.tiff')):
-                from tifffile import imread
-                return(imread(var_path))                
-            if(var_path.suffix == '.torch'):      
-                from torch import load as torch_load 
-                return(torch_load(var_path))
-            try:
-                img = imread(var_path)
-                return(img)
-            except:
-                pass
-    
-    def get_stack_of_files(self, 
-        var_name = None, flist = [], suffix = '*',
-        return_data = False, return_flist = True) -> tuple:
-        
-        """ Get list or data of all files in a directory
-        
-            This function gives the list of paths of all files in a directory
-            for a single variable. 
-
-            :param var_name:
-                The directory or variable name to look for the files
-            :type var_name: str
-            
-            :param flist:
-                list of Paths, if data is returned, this flist input can limit 
-                the data requested to this list.
-            :type flist: list
-            
-            :param suffix:
-                the suffix of files to look for, e.g. 'txt'
-            :type siffix: str
-            
-            :param return_data: 
-                    with flist you can limit the data that is returned.
-                    Otherwise the data for all files in the directory will be
-                    returned
-            :param return_flist
-                    Maybe you are only intrested in the flist.
-                    
-            :return:
-                It returns a tuple, (dataset, flist),
-                dataset will be a numpy array in case all files produce same
-                shape numpy arrays.
-                flist is type pathlib.Path
-            :rtype: tuple
-            
-            
-        """
-        suffix = suffix.strip('.')
-        if not flist:
-            assert var_name is not None, \
-                ' The file list is empty. Please provide the ' \
-                + 'variable name or a non-empty file list.'
-            var_dir = self.log_dir / var_name
-            if(var_dir.is_dir()):
-                var_fname = None
-                flist = list(var_dir.glob(f'*.{suffix}'))
-            else:
-                var_fname = var_dir.name
-                var_dir = var_dir.parent
-                patt = f'{var_fname}*.{suffix}'
-                patt = patt.replace('**', '*')
-                flist = list(var_dir.glob(patt))
-        if flist:
-            flist.sort()
-            n_files = len(flist)
-            if((not return_data) & return_flist):
-                return(flist)
-            data_type = None
-            if(data_type is None):
-                try:
-                    fdata = np.load(flist[0])
-                    data_type = 'numpy'
-                except:
-                    pass
-            if(data_type is None):
-                try:
-                    fdata = imread(flist[0])
-                    data_type = 'image'
-                except:
-                    pass
-            if(data_type is not None):
-                dataset = np.zeros((n_files, ) + fdata.shape, 
-                                   dtype=fdata.dtype)
-                for fcnt, fpath in enumerate(flist):
-                    if(data_type == 'numpy'):
-                        dataset[fcnt] = np.load(fpath)
-                    elif(data_type == 'image'):
-                        dataset[fcnt] = imread(fpath)
-                # self.log_text(None, f'shape is: {dataset.shape}')
-                if(return_flist):
-                    return(dataset, flist)
-                else:
-                    return(dataset)
-            else:
-                # self.log_text(None, f'File {flist[0].name} cannot be opened'\
-                #           + r' by np.load() or plt.imread()')
-                pass
-            
-    def get_common_files(self, var_name_A, var_name_B):
-        """ get common files in two directories
-        
-            It happens often in ML that there are two directories, A and B,
-            and we are interested to get the flist in both that is common 
-            between them. returns a tuple of two lists of files.
-            
-            :param var_name_A:
-                directory A name
-            :param var_name_B:
-                directory B name
-        """
-        flist_A = self.get_stack_of_files(
-            var_name_A, return_data = False, return_flist = True)
-        flist_B = self.get_stack_of_files(
-            var_name_B, return_data = False, return_flist = True)
-        
-        suffix_A = flist_A[0].suffix
-        suffix_B = flist_B[0].suffix 
-        parent_A = flist_A[0].parent
-        parent_B = flist_B[0].parent
-        
-        fstems_A = [_fst.stem for _fst in flist_A]
-        fstems_B = [_fst.stem for _fst in flist_B]
-        
-        fstems_A_set = set(fstems_A)
-        fstems_B_set = set(fstems_B)
-        common_stems = list(fstems_A_set.intersection(fstems_B_set))
-
-        flist_A_new = [parent_A / (common_stem + suffix_A) \
-                          for common_stem in common_stems]
-        flist_B_new = [parent_B / (common_stem + suffix_B) \
-                          for common_stem in common_stems]
-
-        return(flist_A_new, flist_B_new)
-    
     def replace_time_with_index(self, var_name):
         """ index in file names
             lognflow uses time stamps to make new log files for a variable.
@@ -1633,14 +1487,10 @@ class lognflow:
 
     def __call__(self, *args, **kwargs):
         """calling the object
-        
             In the case of the following code::
-            
                 logger = lognflow()
                 logger('Hello lognflow')
-            
             The text (str(...)) will be passed to the main log text file.
-        
         """
         self.log_text(None, *args, **kwargs)